+++ src/common/utility/StringHelpers.h
@@ -127,7 +127,7 @@ namespace StringHelpers
     std::string UTILITY_API Plural(int, const std::string &noun);
     std::string UTILITY_API HumanReadableList(const std::vector<std::string>&);
     bool UTILITY_API IsPureASCII(const std::string &txt);
-    bool UTILITY_API IsPureASCII(const char *const txt, int length);
+    bool UTILITY_API IsPureASCII(const char *const txt, size_t length);
     bool UTILITY_API CaseInsenstiveEqual(const std::string &str_a,
                                          const std::string &str_b);
